Fixed the 'this' problem
It turns out that It wasn't referencing the right object because the
function was being access by another function and not by the object
itself. Now, the class Engine.Time.Event accept a parent by the
parameters and use this to reference the function. But that's optional.
If the user doesn't pass the parent, it just executes the function as it
did before.

Now there's another problem. A Entity is being generated somewhere in
the code and it fails to access the property 'position' of this object.