Fixed the unexpected  Entity
The way I've implemented inheritance was the problem. The algorithm had
to create an Entity in order to extend from it. But I used a worked
around for this and the code is more consistent.

The string on the right, at the bottom of the FPS graph now accept an
array. So multiple information can be very easily added there.